Reorganised darts.models files
