+++ darts/models/filtering/moving_average.py
@@ -3,8 +3,8 @@ Moving Average
 --------------
 """
 
-from .filtering_model import FilteringModel
-from ..timeseries import TimeSeries
+from darts.models.filtering.filtering_model import FilteringModel
+from darts.timeseries import TimeSeries
 
 
 class MovingAverage(FilteringModel):
