+++ darts/models/forecasting/arima.py
@@ -14,9 +14,9 @@ from statsmodels.tsa.arima.model import ARIMA as staARIMA
 from typing import Optional, Tuple
 import numpy as np
 
-from .forecasting_model import DualCovariatesForecastingModel
-from ..timeseries import TimeSeries
-from ..logging import get_logger
+from darts.models.forecasting.forecasting_model import DualCovariatesForecastingModel
+from darts.timeseries import TimeSeries
+from darts.logging import get_logger
 logger = get_logger(__name__)
