+++ darts/models/forecasting/auto_arima.py
@@ -6,9 +6,9 @@ Auto-ARIMA
 from pmdarima import AutoARIMA as PmdAutoARIMA
 from typing import Optional
 
-from .forecasting_model import DualCovariatesForecastingModel
-from ..timeseries import TimeSeries
-from ..logging import get_logger, raise_if
+from darts.models.forecasting.forecasting_model import DualCovariatesForecastingModel
+from darts.timeseries import TimeSeries
+from darts.logging import get_logger, raise_if
 
 logger = get_logger(__name__)
